+++ samples/counter_collection/callback_client.cpp
@@ -62,7 +62,7 @@ namespace
 rocprofiler_context_id_t&
 get_client_ctx()
 {
-    static rocprofiler_context_id_t ctx;
+    static rocprofiler_context_id_t ctx{0};
     return ctx;
 }