+++ samples/counter_collection/print_functional_counters.cpp
@@ -45,7 +45,7 @@ namespace
 rocprofiler_context_id_t&
 get_client_ctx()
 {
-    static rocprofiler_context_id_t ctx;
+    static rocprofiler_context_id_t ctx{0};
     return ctx;
 }