+++ source/include/rocprofiler-sdk/registration.h
@@ -153,7 +153,7 @@ rocprofiler_is_finalized(int* status) ROCPROFILER_API ROCPROFILER_NONNULL(1);
  *
  *      assert(*static_cast<int*>(tool_data) == 1234 && "tool_data is wrong");
  *
- *      rocprofiler_context_id_t ctx;
+ *      rocprofiler_context_id_t ctx{0};
  *      rocprofiler_create_context(&ctx);
  *
  *      if(int valid_ctx = 0;