+++ source/lib/rocprofiler-sdk/context.cpp
@@ -41,6 +41,7 @@ extern "C" {
 rocprofiler_status_t
 rocprofiler_create_context(rocprofiler_context_id_t* context_id)
 {
+    if(context_id->handle != 0) return ROCPROFILER_STATUS_ERROR_CONTEXT_ID_NOT_ZERO;
     // context already registered
     if(rocprofiler::context::get_registered_context(*context_id))
         return ROCPROFILER_STATUS_ERROR_CONTEXT_INVALID;